Extract resolver class for page cursors
